Backend/Server/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
Pillow==8.4.0
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.11.0
